test: fold test_base64 into test_xapi.c
